+++ tests/BareEnum.php
@@ -1,0 +1,11 @@
+<?php
+namespace Wikimedia\JsonCodec\Tests;
+
+/**
+ * Sample UnitEnum using no backing values.
+ */
+enum BareEnum {
+	case ONE;
+	case TWO;
+	case THREE;
+}